wip: fix rotation transform

+++ Source/Editor/OdysseyPainterEditor/Public/Tools/RasterTransformTool/OdysseyPainterEditorRasterTransformTool.h
@@ -108,5 +108,5 @@ private:
     FVector2D mMouseLastReferencePoint;
     ::ULIS::FRectI mInitialSelectionBoundingBox;
 
-    double mRotation;
+    int mRotation;
 };
